Create  Sort Colors
We have take three pointer low ,mid and high . low is placed after the zeros and high is placed before the 2. Here mid is used for traversing in the array if in the array there is 0 found then swap with low and increment low and mid if 1 found then simply increment mid and if 2 found then swap with mid and just decrementing high.