Create Maximum Points You Can Obtain from Cards
First of all we declare ans for keep track of final sum. temp for current sum we add the card points from left to right and store them on the temp. and check the ans between the temp who is max. after that we decrement the temp  from right to left and add the card points from right to left by n-k+i and check the answer by max. and finally return the final ans.