Create Spiral Matrix
if right and bottom is greater or equal to left and top first of all we traverse from left to right and increment top. then traverse from top to bottom and decrement right. if bottom is greater or equal top then  traverse from right to left and decrement bottom then if right is greater or equal to left then we traverse from bottom to top.