Create Word Search
The search function is a recursive helper function used to perform the actual search for the word on the board. It takes in the board (b), the word (w), and the current position (r and c) on the board to start the search.

The base case of the recursion is when the word is empty, indicating that all characters have been matched successfully. In this case, the function returns true.

The function first stores the current character at the current position (val) and then replaces it with '*' to mark it as visited.

It checks the adjacent cells in four directions (up, left, down, right) to see if the next character in the word matches. If a match is found, the function calls itself recursively with the updated word (excluding the first character) and the new position.

The function uses a boolean variable dis to keep track of whether a match was found in any of the adjacent cells.

After exploring all possible directions, the function restores the original character at the current position