Create Max Value of Equation
the code finds the maximum value of the equation points[i][0] + points[i][1] + x[j] - y[j], where x[j] and y[j] are two points within the given set of points such that the x-coordinate difference between them is less than or equal to k. The priority queue is used to keep track of the potential points within the constraint, and the maximum value of the equation is updated whenever a new point is considered.